+++ src/commonMain/kotlin/csense/kotlin/datastructures/CsenseExperimentalDataStructure.kt
@@ -1,0 +1,4 @@
+package csense.kotlin.datastructures
+
+@RequiresOptIn("This is subject to change or be deleted. (experimental construct)", RequiresOptIn.Level.WARNING)
+public annotation class CsenseExperimentalDataStructure