+++ src/commonMain/kotlin/csense/kotlin/datastructures/values/state/SwapBoolean.kt
@@ -1,0 +1,8 @@
+package csense.kotlin.datastructures.values.state
+
+public class SwapBoolean(
+    initialValue: Boolean
+) : SwapValue<Boolean>(
+    initial = initialValue,
+    secondaryState = !initialValue
+)