+++ src/commonMain/kotlin/csense/kotlin/datastructures/values/state/SwapValue.kt
@@ -1,0 +1,19 @@
+package csense.kotlin.datastructures.values.state
+
+public open class SwapValue<T>(
+    initial: T,
+    secondaryState: T
+) : Swappable<T> {
+
+    override val value: T
+        get() = states.value
+
+    private val states = CircularStates(
+        states = listOf(
+            initial,
+            secondaryState
+        )
+    )
+
+    public override fun swap(): Unit = states.next()
+}