+++ changelog.md
@@ -1,10 +1,22 @@
 # Changelog
 
+## 0.0.50
+
+- Added 
+  - map.computeDiff computation
+  - BiMap + MutableBiMap
+  - CircularStates
+  - SwapBoolean
+  - SwapValue
+
 ## 0.0.42
+
 - more extensions and more tests for LockableValue
 
 ## 0.0.41
+
 - removed JMH from jvm target
 
 ## 0.0.40
+
 - first version (meant to start with csense kotlin 0.0.40 )